Rename ruleset and improve code
